requirements.txt
Flask==2.1.0
Flask-SocketIO==5.1.1
python-dotenv==0.19.2